[Chore] move `Strings.errorToMessage` to `panic.ts`
