+++ src/base/common/utilities/string.ts
@@ -71,44 +71,7 @@ export namespace Strings {
         return result;
     }
     
-    /**
-     * @description Try to convert an error to a human readable message in string.
-     * @param error The given error.
-     * @param verbose If output the stack trace.
-     * @returns A string formated error message.
-     *
-     * @note This function never throws.
-     */
-    export function errorToMessage(error: any, verbose: boolean = true): string {
-        if (!error) {
-            return UNKNOWN_MESSAGE;
-        }
 
-        if (Array.isArray(error)) {
-            const errors = Arrays.coalesce(error);
-            const firstErrorMessage = errorToMessage(errors[0], verbose);
-
-            if (errors.length > 1) {
-                return Strings.format('{0}, ({1} more errors in total)', [firstErrorMessage, errors.length - 1]);
-            }
-
-            return firstErrorMessage;
-        }
-
-        if (typeof error === 'string') {
-            return error;
-        }
-
-        if (error.stack && verbose) {
-            return Strings.format('{0} (stack trace - {1})', [error.message || UNKNOWN_MESSAGE, __stackToMessage(error.stack)]);
-        }
-
-        if (error.message) {
-            return error.message;
-        }
-
-        return `${UNKNOWN_MESSAGE}: ${JSON.stringify(error)}`;
-    }
 }
 
 /**
@@ -116,11 +79,3 @@ export namespace Strings {
  */
 export type UUID = string;
 
-const UNKNOWN_MESSAGE = 'An unknown error occured. Please consult the log for more details.';
-function __stackToMessage(stack: any): string {
-    if (Array.isArray(stack)) {
-        return stack.join('\n');
-    } else {
-        return stack;
-    }
-}
