+++ src/platform/files/common/fileService.ts
@@ -11,8 +11,8 @@ import { Blocker } from "src/base/common/utilities/async";
 import { Mutable, Pair } from "src/base/common/utilities/type";
 import { IRawResourceChangeEvents } from "src/platform/files/common/watcher";
 import { IService, createService } from "src/platform/instantiation/common/decorator";
-import { Strings } from "src/base/common/utilities/string";
 import { noop } from "src/base/common/performance";
+import { errorToMessage } from "src/base/common/utilities/panic";
 
 export const IFileService = createService<IFileService>('file-service');
 
@@ -195,7 +195,7 @@ export class FileService extends Disposable implements IFileService {
         .toAsync()
         .andThen(async provider => <const>[await provider.stat(uri), provider])
         .andThen(([stat, provider]) => this.__resolveStat(uri, provider, stat, opts))
-        .mapErr((error) => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)));
+        .mapErr((error) => new FileOperationError(errorToMessage(error), getFileErrorCode(error)));
     }
 
     public readFile(uri: URI, opts?: IReadFileOptions): AsyncResult<DataBuffer, FileOperationError> {
@@ -253,7 +253,7 @@ export class FileService extends Disposable implements IFileService {
             .andThen(provider => provider.stat(uri))
             .andThen(() => ok(true))
             .orElse(() => ok(false))
-            .mapErr(error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)));
+            .mapErr(error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)));
     }
 
     public createFile(
@@ -270,7 +270,7 @@ export class FileService extends Disposable implements IFileService {
         return this.__getProvider(uri)
             .toAsync()
             .andThen(provider => provider.readdir(uri))
-            .mapErr(error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)));
+            .mapErr(error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)));
     }
 
     public createDir(uri: URI): AsyncResult<void, FileOperationError> {
@@ -318,7 +318,7 @@ export class FileService extends Disposable implements IFileService {
     public delete(uri: URI, opts?: IDeleteFileOptions): AsyncResult<void, FileOperationError> {
         return this.__validateDelete(uri, opts)
             .andThen(provider => provider.delete(uri, { useTrash: !!opts?.useTrash, recursive: !!opts?.recursive }))
-            .orElse(error => err(new FileOperationError(`unable to delete uri: '${URI.toFsPath(uri)}'. Reason: ${Strings.errorToMessage(error)}`, getFileErrorCode(error))));
+            .orElse(error => err(new FileOperationError(`unable to delete uri: '${URI.toFsPath(uri)}'. Reason: ${errorToMessage(error)}`, getFileErrorCode(error))));
     }
 
     public watch(uri: URI, opts?: IWatchOptions): AsyncResult<IDisposable, FileOperationError> {
@@ -337,7 +337,7 @@ export class FileService extends Disposable implements IFileService {
 
         const result = Result.fromThrowable(
             () => provider.watch(uri, opts),
-            error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+            error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
         );
         
         if (result.isErr()) {
@@ -346,7 +346,7 @@ export class FileService extends Disposable implements IFileService {
 
         return Result.fromPromise(
             () => result.unwrap(),
-            error => new FileOperationError(`Cannot watch at target: '${URI.toString(uri)}'. Reason: ${Strings.errorToMessage(error)}`, FileOperationErrorType.UNKNOWN))
+            error => new FileOperationError(`Cannot watch at target: '${URI.toString(uri)}'. Reason: ${errorToMessage(error)}`, FileOperationErrorType.UNKNOWN))
         .andThen(disposable => {
             this._activeWatchers.set(uri, disposable);
             return ok(disposable);
@@ -484,7 +484,7 @@ export class FileService extends Disposable implements IFileService {
             // write through a provider
             return Result.fromPromise(
                 async () => { await provider.writeFile(uri, buffer.buffer, { create: opts?.create ?? false, overwrite: opts?.overwrite ?? false, unlock: opts?.unlock ?? false }); },
-                error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+                error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
             );    
         })());
     }
@@ -498,7 +498,7 @@ export class FileService extends Disposable implements IFileService {
     {
         return Result.fromPromise(
             async () => await provider.open(uri, { create: true, unlock: opts?.unlock ?? false }),
-            error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+            error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
         )
         .andThen(fd => {
             const blocker = new Blocker<void>();
@@ -531,7 +531,7 @@ export class FileService extends Disposable implements IFileService {
 
             return Result.fromPromise(
                 () => blocker.waiting().finally(() => provider.close(fd)),
-                error => new FileOperationError(`unable to write the file buffered ${URI.toFsPath(uri)}. Reason: ${Strings.errorToMessage(error)}`, getFileErrorCode(error)),
+                error => new FileOperationError(`unable to write the file buffered ${URI.toFsPath(uri)}. Reason: ${errorToMessage(error)}`, getFileErrorCode(error)),
             );
         });
     }
@@ -592,7 +592,7 @@ export class FileService extends Disposable implements IFileService {
 
             const res = await Result.fromPromise(
                 async () => provider.mkdir(path),
-                error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+                error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
             );
 
             if (res.isErr()) {
@@ -630,7 +630,7 @@ export class FileService extends Disposable implements IFileService {
         if (stat.type === FileType.DIRECTORY && opts && (opts.resolveChildren || opts.resolveChildrenRecursive)) {
             const dirResult = await Result.fromPromise(
                 () => provider.readdir(uri),
-                error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+                error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
             );
             
             if (dirResult.isErr()) {
@@ -723,7 +723,7 @@ export class FileService extends Disposable implements IFileService {
             } 
             return Result.fromPromise(
                 () => fromProvider.copy(from, to, { overwrite: !!overwrite }),
-                error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+                error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
             );
         });
     }
@@ -793,7 +793,7 @@ export class FileService extends Disposable implements IFileService {
     {
         return Result.fromPromise(
             () => provider.stat(uri),
-            error => new FileOperationError(Strings.errorToMessage(error), getFileErrorCode(error)),
+            error => new FileOperationError(errorToMessage(error), getFileErrorCode(error)),
         )
         .andThen(stat => {
             if (!stat) {
